Consolidate timeout thumbnail effect
